Add files via upload
The file titled "800_5dim.csv" contains 5-dimensional condition monitoring data, generated by a simulated industrial fleet comprising of 800 assets. The file contains 6 columns, where the first column denotes the asset id. The following five columns resemble the simulated sensors or any other source of condition data. Each asset's data comprises of 500 data points, which are sampled from a stationary Gaussian. Depending on the experimental case, a subset of these data points shall be used. The overall fleet comprises of four asset clusters. In the dataset shared here, the first 200 assets (asset id 1 to 200) belong to cluster 1, the following 200 assets (asset id 201 to 400) belong to cluster 2, and so on.

Extended description about the simulation process is explained in <cite the DCE paper here after it is published>.